(maint) [WIP] [DO NOT MERGE] Refactor Puppet::Type
Factoring in the original sense of tearing it apart into its indivisible
parts. This is mostly an exploration of which parts are there even. It's
also not done yet, but I wanted to upload what I have for now, to gather
feedback from interested parties.

Besides the academic exercise of working with the code to understand it,
I hope that breaking out the chunks helps to better map out the class'
responsibilities.

For example in querying_attributes_original, the gaps and inconsistencies
of the existing APIs becomes very apparent.
