+++ lib/puppet/type_guts/attribute_definitions.rb
@@ -1,0 +1,463 @@
+module Puppet
+  class Type
+    newmetaparam(:noop) do
+      desc "Whether to apply this resource in noop mode.
+
+      When applying a resource in noop mode, Puppet will check whether it is in sync,
+      like it does when running normally. However, if a resource attribute is not in
+      the desired state (as declared in the catalog), Puppet will take no
+      action, and will instead report the changes it _would_ have made. These
+      simulated changes will appear in the report sent to the puppet master, or
+      be shown on the console if running puppet agent or puppet apply in the
+      foreground. The simulated changes will not send refresh events to any
+      subscribing or notified resources, although Puppet will log that a refresh
+      event _would_ have been sent.
+
+      **Important note:**
+      [The `noop` setting](https://docs.puppetlabs.com/puppet/latest/reference/configuration.html#noop)
+      allows you to globally enable or disable noop mode, but it will _not_ override
+      the `noop` metaparameter on individual resources. That is, the value of the
+      global `noop` setting will _only_ affect resources that do not have an explicit
+      value set for their `noop` attribute."
+
+      newvalues(:true, :false)
+      munge do |value|
+        case value
+          when true, :true, "true"; @resource.noop = true
+          when false, :false, "false"; @resource.noop = false
+        end
+      end
+    end
+
+    newmetaparam(:schedule) do
+      desc "A schedule to govern when Puppet is allowed to manage this resource.
+      The value of this metaparameter must be the `name` of a `schedule`
+      resource. This means you must declare a schedule resource, then
+      refer to it by name; see
+      [the docs for the `schedule` type](https://docs.puppetlabs.com/puppet/latest/reference/type.html#schedule)
+      for more info.
+
+          schedule { 'everyday':
+            period => daily,
+            range  => \"2-4\"
+          }
+
+          exec { \"/usr/bin/apt-get update\":
+            schedule => 'everyday'
+          }
+
+      Note that you can declare the schedule resource anywhere in your
+      manifests, as long as it ends up in the final compiled catalog."
+    end
+
+    newmetaparam(:audit) do
+      desc "Marks a subset of this resource's unmanaged attributes for auditing. Accepts an
+      attribute name, an array of attribute names, or `all`.
+
+      Auditing a resource attribute has two effects: First, whenever a catalog
+      is applied with puppet apply or puppet agent, Puppet will check whether
+      that attribute of the resource has been modified, comparing its current
+      value to the previous run; any change will be logged alongside any actions
+      performed by Puppet while applying the catalog.
+
+      Secondly, marking a resource attribute for auditing will include that
+      attribute in inspection reports generated by puppet inspect; see the
+      puppet inspect documentation for more details.
+
+      Managed attributes for a resource can also be audited, but note that
+      changes made by Puppet will be logged as additional modifications. (I.e.
+      if a user manually edits a file whose contents are audited and managed,
+      puppet agent's next two runs will both log an audit notice: the first run
+      will log the user's edit and then revert the file to the desired state,
+      and the second run will log the edit made by Puppet.)"
+
+      validate do |list|
+        list = Array(list).collect {|p| p.to_sym}
+        unless list == [:all]
+          list.each do |param|
+            next if @resource.class.validattr?(param)
+            fail "Cannot audit #{param}: not a valid attribute for #{resource}"
+          end
+        end
+      end
+
+      munge do |args|
+        properties_to_audit(args).each do |param|
+          next unless resource.class.validproperty?(param)
+          resource.newattr(param)
+        end
+      end
+
+      def all_properties
+        resource.class.properties.find_all do |property|
+          resource.provider.nil? or resource.provider.class.supports_parameter?(property)
+        end.collect do |property|
+          property.name
+        end
+      end
+
+      def properties_to_audit(list)
+        if !list.kind_of?(Array) && list.to_sym == :all
+          list = all_properties
+        else
+          list = Array(list).collect { |p| p.to_sym }
+        end
+      end
+    end
+
+    newmetaparam(:loglevel) do
+      desc "Sets the level that information will be logged.
+      The log levels have the biggest impact when logs are sent to
+      syslog (which is currently the default).
+
+      The order of the log levels, in decreasing priority, is:
+
+      * `crit`
+      * `emerg`
+      * `alert`
+      * `err`
+      * `warning`
+      * `notice`
+      * `info` / `verbose`
+      * `debug`
+      "
+      defaultto :notice
+
+      newvalues(*Puppet::Util::Log.levels)
+      newvalues(:verbose)
+
+      munge do |loglevel|
+        val = super(loglevel)
+        if val == :verbose
+          val = :info
+        end
+        val
+      end
+    end
+
+    newmetaparam(:alias) do
+      desc %q{Creates an alias for the resource.  Puppet uses this internally when you
+      provide a symbolic title and an explicit namevar value:
+
+          file { 'sshdconfig':
+            path => $operatingsystem ? {
+              solaris => '/usr/local/etc/ssh/sshd_config',
+              default => '/etc/ssh/sshd_config',
+            },
+            source => '...'
+          }
+
+          service { 'sshd':
+            subscribe => File['sshdconfig'],
+          }
+
+      When you use this feature, the parser sets `sshdconfig` as the title,
+      and the library sets that as an alias for the file so the dependency
+      lookup in `Service['sshd']` works.  You can use this metaparameter yourself,
+      but note that aliases generally only work for creating relationships; anything
+      else that refers to an existing resource (such as amending or overriding
+      resource attributes in an inherited class) must use the resource's exact
+      title. For example, the following code will not work:
+
+          file { '/etc/ssh/sshd_config':
+            owner => root,
+            group => root,
+            alias => 'sshdconfig',
+          }
+
+          File['sshdconfig'] {
+            mode => '0644',
+          }
+
+      There's no way here for the Puppet parser to know that these two stanzas
+      should be affecting the same file.
+
+      }
+
+      munge do |aliases|
+        aliases = [aliases] unless aliases.is_a?(Array)
+
+        raise(ArgumentError, "Cannot add aliases without a catalog") unless @resource.catalog
+
+        aliases.each do |other|
+          if obj = @resource.catalog.resource(@resource.class.name, other)
+            unless obj.object_id == @resource.object_id
+              self.fail("#{@resource.title} can not create alias #{other}: object already exists")
+            end
+            next
+          end
+
+          # Newschool, add it to the catalog.
+          @resource.catalog.alias(@resource, other)
+        end
+      end
+    end
+
+    newmetaparam(:tag) do
+      desc "Add the specified tags to the associated resource.  While all resources
+      are automatically tagged with as much information as possible
+      (e.g., each class and definition containing the resource), it can
+      be useful to add your own tags to a given resource.
+
+      Multiple tags can be specified as an array:
+
+          file {'/etc/hosts':
+            ensure => file,
+            source => 'puppet:///modules/site/hosts',
+            mode   => '0644',
+            tag    => ['bootstrap', 'minimumrun', 'mediumrun'],
+          }
+
+      Tags are useful for things like applying a subset of a host's configuration
+      with [the `tags` setting](/puppet/latest/reference/configuration.html#tags)
+      (e.g. `puppet agent --test --tags bootstrap`)."
+
+      munge do |tags|
+        tags = [tags] unless tags.is_a? Array
+
+        tags.each do |tag|
+          @resource.tag(tag)
+        end
+      end
+    end
+
+    # RelationshipMetaparam is an implementation supporting the meta-parameters `:require`, `:subscribe`,
+    # `:notify`, and `:before`.
+    #
+    #
+    class RelationshipMetaparam < Puppet::Parameter
+      class << self
+        attr_accessor :direction, :events, :callback, :subclasses
+      end
+
+      @subclasses = []
+
+      def self.inherited(sub)
+        @subclasses << sub
+      end
+
+      # @return [Array<Puppet::Resource>] turns attribute value(s) into list of resources
+      def munge(references)
+        references = [references] unless references.is_a?(Array)
+        references.collect do |ref|
+          if ref.is_a?(Puppet::Resource)
+            ref
+          else
+            Puppet::Resource.new(ref)
+          end
+        end
+      end
+
+      # Checks each reference to assert that what it references exists in the catalog.
+      #
+      # @raise [???fail] if the referenced resource can not be found
+      # @return [void]
+      def validate_relationship
+        @value.each do |ref|
+          unless @resource.catalog.resource(ref.to_s)
+            description = self.class.direction == :in ? "dependency" : "dependent"
+            fail ResourceError, "Could not find #{description} #{ref} for #{resource.ref}"
+          end
+        end
+      end
+
+      # Creates edges for all relationships.
+      # The `:in` relationships are specified by the event-receivers, and `:out`
+      # relationships are specified by the event generator.
+      # @todo references to "event-receivers" and "event generator" means in this context - are those just
+      #   the resources at the two ends of the relationship?
+      # This way 'source' and 'target' are consistent terms in both edges
+      # and events, i.e. an event targets edges whose source matches
+      # the event's source. The direction of the relationship determines
+      # which resource is applied first and which resource is considered
+      # to be the event generator.
+      # @return [Array<Puppet::Relationship>]
+      # @raise [???fail] when a reference can not be resolved
+      #
+      def to_edges
+        @value.collect do |reference|
+          reference.catalog = resource.catalog
+
+          # Either of the two retrieval attempts could have returned
+          # nil.
+          unless related_resource = reference.resolve
+            self.fail "Could not retrieve dependency '#{reference}' of #{@resource.ref}"
+          end
+
+          # Are we requiring them, or vice versa?  See the method docs
+          # for further info on this.
+          if self.class.direction == :in
+            source = related_resource
+            target = @resource
+          else
+            source = @resource
+            target = related_resource
+          end
+
+          if method = self.class.callback
+            subargs = {
+                :event => self.class.events,
+                :callback => method
+            }
+            self.debug { "subscribes to #{related_resource.ref}" }
+          else
+            # If there's no callback, there's no point in even adding
+            # a label.
+            subargs = nil
+            self.debug { "subscribes to #{related_resource.ref}" }
+          end
+
+          Puppet::Relationship.new(source, target, subargs)
+        end
+      end
+    end
+
+    # Note that the order in which the relationships params is defined
+    # matters.  The labeled params (notify and subscribe) must be later,
+    # so that if both params are used, those ones win.  It's a hackish
+    # solution, but it works.
+
+    newmetaparam(:require, :parent => RelationshipMetaparam, :attributes => {:direction => :in, :events => :NONE}) do
+      desc "One or more resources that this resource depends on, expressed as
+      [resource references](https://docs.puppetlabs.com/puppet/latest/reference/lang_data_resource_reference.html).
+      Multiple resources can be specified as an array of references. When this
+      attribute is present:
+
+      * The required resource(s) will be applied **before** this resource.
+
+      This is one of the four relationship metaparameters, along with
+      `before`, `notify`, and `subscribe`. For more context, including the
+      alternate chaining arrow (`->` and `~>`) syntax, see
+      [the language page on relationships](https://docs.puppetlabs.com/puppet/latest/reference/lang_relationships.html)."
+    end
+
+    newmetaparam(:subscribe, :parent => RelationshipMetaparam, :attributes => {:direction => :in, :events => :ALL_EVENTS, :callback => :refresh}) do
+      desc "One or more resources that this resource depends on, expressed as
+      [resource references](https://docs.puppetlabs.com/puppet/latest/reference/lang_data_resource_reference.html).
+      Multiple resources can be specified as an array of references. When this
+      attribute is present:
+
+      * The subscribed resource(s) will be applied _before_ this resource.
+      * If Puppet makes changes to any of the subscribed resources, it will cause
+        this resource to _refresh._ (Refresh behavior varies by resource
+        type: services will restart, mounts will unmount and re-mount, etc. Not
+        all types can refresh.)
+
+      This is one of the four relationship metaparameters, along with
+      `before`, `require`, and `notify`. For more context, including the
+      alternate chaining arrow (`->` and `~>`) syntax, see
+      [the language page on relationships](https://docs.puppetlabs.com/puppet/latest/reference/lang_relationships.html)."
+    end
+
+    newmetaparam(:before, :parent => RelationshipMetaparam, :attributes => {:direction => :out, :events => :NONE}) do
+      desc "One or more resources that depend on this resource, expressed as
+      [resource references](https://docs.puppetlabs.com/puppet/latest/reference/lang_data_resource_reference.html).
+      Multiple resources can be specified as an array of references. When this
+      attribute is present:
+
+      * This resource will be applied _before_ the dependent resource(s).
+
+      This is one of the four relationship metaparameters, along with
+      `require`, `notify`, and `subscribe`. For more context, including the
+      alternate chaining arrow (`->` and `~>`) syntax, see
+      [the language page on relationships](https://docs.puppetlabs.com/puppet/latest/reference/lang_relationships.html)."
+    end
+
+    newmetaparam(:notify, :parent => RelationshipMetaparam, :attributes => {:direction => :out, :events => :ALL_EVENTS, :callback => :refresh}) do
+      desc "One or more resources that depend on this resource, expressed as
+      [resource references](https://docs.puppetlabs.com/puppet/latest/reference/lang_data_resource_reference.html).
+      Multiple resources can be specified as an array of references. When this
+      attribute is present:
+
+      * This resource will be applied _before_ the notified resource(s).
+      * If Puppet makes changes to this resource, it will cause all of the
+        notified resources to _refresh._ (Refresh behavior varies by resource
+        type: services will restart, mounts will unmount and re-mount, etc. Not
+        all types can refresh.)
+
+      This is one of the four relationship metaparameters, along with
+      `before`, `require`, and `subscribe`. For more context, including the
+      alternate chaining arrow (`->` and `~>`) syntax, see
+      [the language page on relationships](https://docs.puppetlabs.com/puppet/latest/reference/lang_relationships.html)."
+    end
+
+    newmetaparam(:stage) do
+      desc %{Which run stage this class should reside in.
+
+      **Note: This metaparameter can only be used on classes,** and only when
+      declaring them with the resource-like syntax. It cannot be used on normal
+      resources or on classes declared with `include`.
+
+      By default, all classes are declared in the `main` stage. To assign a class
+      to a different stage, you must:
+
+      * Declare the new stage as a [`stage` resource](https://docs.puppetlabs.com/puppet/latest/reference/type.html#stage).
+      * Declare an order relationship between the new stage and the `main` stage.
+      * Use the resource-like syntax to declare the class, and set the `stage`
+        metaparameter to the name of the desired stage.
+
+      For example:
+
+          stage { 'pre':
+            before => Stage['main'],
+          }
+
+          class { 'apt-updates':
+            stage => 'pre',
+          }
+    }
+    end
+
+    newmetaparam(:export, :parent => RelationshipMetaparam, :attributes => {:direction => :out, :events => :NONE}) do
+      desc <<EOS
+Export a capability resource.
+
+The value of this parameter must be a reference to a capability resource,
+or an array of such references. Each capability resource referenced here
+will be instantiated in the node catalog and exported to consumers of this
+resource. The title of the capability resource will be the title given in
+the reference, and all other attributes of the resource will be filled
+according to the corresponding produces statement.
+
+It is an error if this metaparameter references resources whose type is not
+a capability type, or of there is no produces clause for the type of the
+current resource and the capability resource mentioned in this parameter.
+
+For example:
+
+define web(..) { .. }
+Web produces Http { .. }
+web { server:
+  export => Http[main_server]
+}
+EOS
+    end
+
+    newmetaparam(:consume, :parent => RelationshipMetaparam, :attributes => {:direction => :in, :events => :NONE}) do
+      desc <<EOS
+Consume a capability resource.
+
+The value of this parameter must be a reference to a capability resource,
+or an array of such references. Each capability resource referenced here
+must have been exported by another resource in the same environment.
+
+The referenced capability resource(s) will be looked up, added to the
+current node catalog, and processed following the underlying consumes
+clause.
+
+It is an error if this metaparameter references resources whose type is not
+a capability type, or of there is no consumes clause for the type of the
+current resource and the capability resource mentioned in this parameter.
+
+For example:
+
+define web(..) { .. }
+Web consumes Sql { .. }
+web { server:
+  consume => Sql[my_db]
+}
+EOS
+    end
+
+  end
+end
